puhkuluettelo steps 11 ja 12

+++ osa2/puhelinluettelo/src/App.jsx
@@ -1,7 +1,9 @@
 import React, { useEffect, useState } from 'react';
 import FilterForm from './components/FilterForm';
+import Notification from './components/Notification';
 import PersonForm from './components/PersonForm';
 import Persons from './components/Persons';
+import './index.css';
 import personsService from './services/PersonsService';
 
 const App = () => {
@@ -9,6 +11,8 @@ const App = () => {
   const [newName, setNewName] = useState('');
   const [newNumber, setNewNumber] = useState('');
   const [filter, setFilter] = useState('');
+  const [errorMessage, setErrorMessage] = useState(null);
+  const [successMessage, setSuccessMessage] = useState(null);
 
   useEffect(() => {
     personsService.getAll().then(initialPersons => {
@@ -30,6 +34,11 @@ const App = () => {
           setPersons(persons.map(person => person.id !== existingPerson.id ? person : returnedPerson));
           setNewName('');
           setNewNumber('');
+          setSuccessMessage(`Updated ${newName}'s number`);
+          setTimeout(() => setSuccessMessage(null), 5000);
+        }).catch(error => {
+          setErrorMessage(`Error updating ${newName}'s number`);
+          setTimeout(() => setErrorMessage(null), 5000);
         });
       }
     } else {
@@ -37,14 +46,27 @@ const App = () => {
         setPersons(persons.concat(returnedPerson));
         setNewName('');
         setNewNumber('');
+        setSuccessMessage(`Added ${newName}`);
+        setTimeout(() => setSuccessMessage(null), 5000);
+      }).catch(error => {
+        setErrorMessage(`Error adding ${newName}`);
+        setTimeout(() => setErrorMessage(null), 5000);
       });
     }
   };
 
   const deletePerson = id => {
-    personsService.remove(id).then(() => {
-      setPersons(persons.filter(person => person.id !== id));
-    });
+    const person = persons.find(p => p.id === id);
+    if (window.confirm(`Delete ${person.name}?`)) {
+      personsService.remove(id).then(() => {
+        setPersons(persons.filter(person => person.id !== id));
+        setSuccessMessage(`Deleted ${person.name}`);
+        setTimeout(() => setSuccessMessage(null), 5000);
+      }).catch(error => {
+        setErrorMessage(`Error deleting ${person.name}`);
+        setTimeout(() => setErrorMessage(null), 5000);
+      });
+    }
   };
 
   const handleNameChange = (event) => {
@@ -66,6 +88,8 @@ const App = () => {
   return (
     <div>
       <h1>Phonebook</h1>
+      <Notification message={errorMessage} type="error" />
+      <Notification message={successMessage} type="success" />
       <FilterForm filter={filter} handleFilterChange={handleFilterChange} />
       <h2>Add a new person</h2>
       <PersonForm
